+++ packages/graphql-codegen-cli/src/utils/watcher.ts
@@ -60,7 +60,17 @@ export const createWatcher = (
   const runWatcher = async (abortSignal: AbortSignal) => {
     const watchDirectory = await findHighestCommonDirectory(allAffirmativePatterns);
 
-    const parcelWatcher = await import('@parcel/watcher');
+    // Try to load the parcel watcher, but don't fail if it's not available.
+    let parcelWatcher: typeof import('@parcel/watcher');
+    try {
+      parcelWatcher = await import('@parcel/watcher');
+    } catch (err) {
+      log(
+        `Failed to import @parcel/watcher due to the following error (to use watch mode, install https://www.npmjs.com/package/@parcel/watcher):\n${err}`
+      );
+      return;
+    }
+
     debugLog(`[Watcher] Parcel watcher loaded...`);
 
     let isShutdown = false;
@@ -74,7 +84,7 @@ export const createWatcher = (
     }, 100);
     emitWatching(watchDirectory);
 
-    const ignored: string[] = [];
+    const ignored: string[] = ['**/.git/**'];
     for (const entry of Object.keys(config.generates).map(filename => ({
       filename,
       config: normalizeOutputParam(config.generates[filename]),
@@ -219,7 +229,11 @@ const findHighestCommonDirectory = async (files: string[]): Promise<string> => {
   // e.g. mm.scan("/**/foo/bar").base -> "/" ; mm.scan("/foo/bar/**/fizz/*.graphql") -> /foo/bar
   const dirPaths = files
     .map(filePath => (isAbsolute(filePath) ? filePath : resolve(filePath)))
-    .map(patterned => mm.scan(patterned).base);
+    // mm.scan doesn't know how to handle Windows \ path separator
+    .map(patterned => patterned.replace(/\\/g, '/'))
+    .map(patterned => mm.scan(patterned).base)
+    // revert the separators to the platform-supported ones
+    .map(base => base.replace(/\//g, sep));
 
   // Return longest common prefix if it's accessible, otherwise process.cwd()
   return (async (maybeValidPath: string) => {