+++ packages/graphql-codegen-cli/tests/codegen.spec.ts
@@ -719,7 +719,7 @@ describe('Codegen Executor', () => {
 
       expect(output.length).toBe(1);
       expect(output[0].content).toBeSimilarStringTo(`export type Scalars = {
-        ID: { input: string | number; output: string; }
+        ID: { input: string; output: string; }
         String: { input: string; output: string; }
         Boolean: { input: boolean; output: boolean; }
         Int: { input: number; output: number; }
@@ -1304,5 +1304,23 @@ describe('Codegen Executor', () => {
       const fileOutput = output.find(file => file.filename === './src/gql/graphql.ts');
       expect(fileOutput.content).toContain('export type BarQuery');
     });
+  });
+
+  it('should not run out of memory when generating very complex types (issue #7720)', async () => {
+    const result = await executeCodegen({
+      schema: ['../../dev-test/gatsby/schema.graphql'],
+      documents: ['../../dev-test/gatsby/fragments.ts'],
+      config: {
+        extractAllFieldsToTypes: true,
+        dedupeOperationSuffix: true,
+      },
+      generates: {
+        'out1.ts': {
+          plugins: ['typescript', 'typescript-operations'],
+        },
+      },
+    });
+    expect(result.length).toBe(1);
+    expect(result[0].content).toContain('export type WpCoreImageBlockForGalleryFragment = ');
   });
 });