+++ packages/plugins/other/visitor-plugin-common/src/client-side-base-visitor.ts
@@ -146,7 +146,7 @@ export interface RawClientSideBasePluginConfig extends RawConfig {
   documentMode?: DocumentMode;
   /**
    * @default true
-   * @description If you are using `documentNode: documentMode | documentNodeImportFragments`, you can set this to `true` to apply document optimizations for your GraphQL document.
+   * @description If you are using `documentMode: documentNode | documentNodeImportFragments`, you can set this to `true` to apply document optimizations for your GraphQL document.
    * This will remove all "loc" and "description" fields from the compiled document, and will remove all empty arrays (such as `directives`, `arguments` and `variableDefinitions`).
    */
   optimizeDocumentNode?: boolean;