+++ packages/presets/client/src/index.ts
@@ -126,6 +126,7 @@ export const preset: Types.OutputPreset<ClientPresetConfig> = {
       skipTypename: options.config.skipTypename,
       arrayInputCoercion: options.config.arrayInputCoercion,
       enumsAsTypes: options.config.enumsAsTypes,
+      futureProofEnums: options.config.futureProofEnums,
       dedupeFragments: options.config.dedupeFragments,
       nonOptionalTypename: options.config.nonOptionalTypename,
       avoidOptionals: options.config.avoidOptionals,
@@ -360,3 +361,5 @@ function createDeferred<T = void>(): Deferred<T> {
   });
   return d;
 }
+
+export { addTypenameSelectionDocumentTransform } from './add-typename-selection-document-transform.js';